requirements.txt
pandas>=0.21.1
requests>=2.18.4
matplotlib>=2.1.1
gdax-python-api>=0.2